% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manage.R
\name{modb_conn_ref}
\alias{modb_conn_ref}
\title{MODB Connection References}
\usage{
modb_conn_ref(conn_id, conn_name, conn_ref, args = NULL)
}
\arguments{
\item{conn_id}{Integer. The ID of the connection returned by modb_connect}

\item{conn_name}{String. The name provided when creating a connection.}

\item{conn_ref}{Expected to be either a conn_id or a conn_name}

\item{args}{list(...) Arguments passed to a calling function in the va scope.}
}
\description{
\code{modb_conn_ref} checks for and/or validates parameters that may be a
connection reference. Connections in MODB may be named on creation or must
be referenced by the unique ID that is returned by the
\link{\code{modb_connect}} method.
}
\seealso{
\link{\code{modb_connect}}
}
