% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manage.R
\name{modb_connect}
\alias{modb_connect}
\title{MODB Connections}
\usage{
modb_connect(
  username,
  password,
  database,
  host = "localhost",
  port = 3306,
  socket = NULL,
  conn_name = NULL
)
}
\arguments{
\item{username}{String. Username to use when connecting to the database.}

\item{password}{String. Password to use when connecting to the database.}

\item{database}{String. The database to use.}

\item{host}{String. The hostname or IP address to connect to.
Either host & port or socket may be used}

\item{port}{Integer. The port number to connect on.}

\item{socket}{String. The socket path to connect to,
e.g. /var/run/mysqld/mysqld.sock}

\item{name}{String. A name for the connection which can be used to identify
the connection later on. Alternatively the id returned
can be used.}
}
\value{
The id of the connection
}
\description{
\code{modb_connect} creates a new database connection and attmepts to open a
connection to a database with the details provided.
}
