% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manage.R
\name{modb_connectionInfo}
\alias{modb_connectionInfo}
\alias{modb_connectionExists}
\alias{modb_connectionId}
\alias{modb_connectionName}
\title{MODB Connections}
\usage{
modb_connectionInfo(stop_on_error = TRUE, ...)

modb_connectionExists(...)

modb_connectionId(conn_name, stop_on_error = TRUE)

modb_connectionName(conn_id, stop_on_error = TRUE)
}
\arguments{
\item{...}{conn_id, conn_name or conn_ref required. See \link{modb_conn_ref}}

\item{conn_name}{String. The name provided when creating a connection.}

\item{conn_id}{Integer. The id of the connection, returned by modb_connect.}
}
\value{
The details of the connection as a list.

TRUE if the connection exists, FALSE if not.

The ID of the connection.

The name of the connection.
}
\description{
Fetches information on a connection such as the last query, insert id and
number of queries run
}
\section{Functions}{
\itemize{
\item \code{modb_connectionExists}: Determines if a connection with the name or
id provided exists

\item \code{modb_connectionId}: Finds a connection id via the name

\item \code{modb_connectionName}: Finds a connection name via the id
}}

