% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manage.R
\name{modb_disconnect}
\alias{modb_disconnect}
\title{MODB Connections}
\usage{
modb_disconnect(...)
}
\arguments{
\item{...}{conn_id, conn_name or conn_ref required. See \link{modb_conn_ref}}
}
\description{
modb_disconnect closes a database connection and cleans up the instance
}
